+++ src/main/java/blackjack/domain/card/CardData.java
@@ -1,4 +1,4 @@
-package blackjack.domain;
+package blackjack.domain.card;
 
 public class CardData {
     public enum Denomination {
