+++ src/main/java/blackjack/domain/card/Deck.java
@@ -1,4 +1,7 @@
-package blackjack.domain;
+package blackjack.domain.card;
+
+import blackjack.domain.card.Card;
+import blackjack.domain.card.CardData;
 
 import java.util.ArrayList;
 import java.util.Collections;
