+++ src/main/java/blackjack/domain/person/Player.java
@@ -1,4 +1,8 @@
-package blackjack.domain;
+package blackjack.domain.person;
+
+import blackjack.domain.card.Card;
+import blackjack.domain.card.Deck;
+import blackjack.domain.GameConstant;
 
 import java.util.ArrayList;
