+++ src/main/java/blackjack/view/OutputView.java
@@ -1,6 +1,9 @@
 package blackjack.view;
 
 import blackjack.domain.*;
+import blackjack.domain.person.Dealer;
+import blackjack.domain.person.Player;
+
 import java.util.ArrayList;
 import java.util.Map;
 import java.util.stream.Collectors;
